requirements.txt
Flask
pyjwt
bcrypt
python-dotenv